opt: minor cleanup to index constraints code
Moving isIndexedVar out of index_constraint_spans, and moving
in IndexColumnInfo.

Release note: None
